+++ config/config.exs
@@ -8,7 +8,3 @@ config :chromic_pdf, debug_protocol: false
 if Mix.env() in [:dev, :test] do
   config :chromic_pdf, dev_pool_size: 1
 end
-
-if Mix.env() in [:dev, :test] do
-  import_config "../examples/phoenix/config.exs"
-end